move separate files into the same component
