+++ src/pages/index.jsx
@@ -1,12 +1,93 @@
 import React from 'react';
 import { Link } from 'gatsby';
+import { StaticImage } from 'gatsby-plugin-image';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faBus, faWalking } from '@fortawesome/free-solid-svg-icons';
 
-import '../styles/css/main.css';
-import '../styles/css/fontawesome-all.min.css';
-import CuImg from '../images/cuimage.jpeg';
-import ContactForm from '../components/ContactForm';
+import '../styles/sass/main.scss';
 import Layout from '../components/Layout';
-import Projects from '../components/Projects';
+
+/**
+ * Component for rendering contact us form
+ * @returns {JSX}
+ */
+function ContactForm() {
+  return (
+    <form
+      name="contactCLTU"
+      method="post"
+      data-netlify="true"
+      className="row gtr-uniform"
+      action="/success/"
+    >
+      <input type="hidden" name="form-name" value="contactCLTU" />
+      <div className="col-6 col-12-xsmall">
+        <label htmlFor="name">
+          Name
+          <input type="text" name="name" id="name" required />
+
+        </label>
+      </div>
+      <div className="col-6 col-12-xsmall">
+        <label htmlFor="email">
+          Email
+          <input type="email" name="email" id="email" required />
+
+        </label>
+      </div>
+      <div className="col-12">
+        <label htmlFor="message">
+          Share your thoughts
+          <textarea name="message" id="message" placeholder="It would be great if..." />
+
+        </label>
+      </div>
+      <div className="col-12">
+        <ul className="actions">
+          <li><button type="submit" className="primary">Submit</button></li>
+        </ul>
+      </div>
+    </form>
+  );
+}
+
+/**
+ * Component for rendering the Projects section
+ * @returns {JSX}
+ */
+function Projects() {
+  const projects = [
+    {
+      icon: faBus,
+      title: 'Bus Rider Comfort',
+      description: 'The stated goal is to foster and facilitate an environment that is friendly for bus riders. To do this, we have decided that adequate seating at bus stops is the best way of improving comfort at a low cost.',
+    },
+    {
+      icon: faWalking,
+      title: 'Reduce Pedestrian Fatalities',
+      description: 'The stated goal is to curb Charlotte\'s status as being one of the worst pedestrian metros in the United States. In order to tackle this, we aim to pacify the top five pedestrian hotspots where pedestrians have been hit and killed.',
+    },
+  ];
+
+  return (
+    <section>
+      <header className="major">
+        <h2>Projects</h2>
+      </header>
+      <div className="features">
+        {projects.map(({ icon, title, description }) => (
+          <article>
+            <FontAwesomeIcon icon={icon} className="icon" />
+            <div className="content">
+              <h3>{title}</h3>
+              <p>{description}</p>
+            </div>
+          </article>
+        ))}
+      </div>
+    </section>
+  );
+}
 
 export default function Home() {
   return (
@@ -17,14 +98,14 @@ export default function Home() {
             <h1>We are the Charlotte Urbanists</h1>
           </header>
           <p>
-            {/* eslint-disable-next-line max-len */}
-            Our Mission is to combat destructive suburban norms by advocating for sustainable, sensible, and equitable urban practices and policies
+            Our Mission is to combat destructive suburban norms by advocating for sustainable,
+            sensible, and equitable urban practices and policies
           </p>
           <ul className="actions">
             <li><Link to="/about" className="button big">Learn More</Link></li>
           </ul>
         </div>
-        <img className="image fit" src={CuImg} alt="An illustration of a public transportation friendly city" />
+        <StaticImage src="../images/cuimage.jpeg" alt="An illustration of a well urbanised city" />
       </section>
       <Projects />
       <section id="contact_form">
